Fulltext Matching Query #ESPP-269

+++ api/src/Elasticsuite/Standard/src/Search/Elasticsearch/Request/Query/MatchQuery.php
@@ -30,7 +30,7 @@ class MatchQuery implements QueryInterface
 
     private ?string $name;
 
-    private int $boost;
+    private float $boost;
 
     private string $queryText;
 
@@ -45,14 +45,14 @@ class MatchQuery implements QueryInterface
      * @param string  $field              Query field
      * @param string  $minimumShouldMatch Minimum should match for the match query
      * @param ?string $name               Query name
-     * @param int     $boost              Query boost
+     * @param float   $boost              Query boost
      */
     public function __construct(
         string $queryText,
         string $field,
         string $minimumShouldMatch = self::DEFAULT_MINIMUM_SHOULD_MATCH,
         ?string $name = null,
-        int $boost = QueryInterface::DEFAULT_BOOST_VALUE
+        float $boost = QueryInterface::DEFAULT_BOOST_VALUE
     ) {
         $this->name = $name;
         $this->queryText = $queryText;
@@ -72,7 +72,7 @@ class MatchQuery implements QueryInterface
     /**
      * {@inheritDoc}
      */
-    public function getBoost(): int
+    public function getBoost(): float
     {
         return $this->boost;
     }
