+++ api/src/Elasticsuite/Standard/src/Search/Elasticsearch/Request/QueryInterface.php
@@ -55,5 +55,5 @@ interface QueryInterface
     /**
      * Get boost.
      */
-    public function getBoost(): int|null;
+    public function getBoost(): float|null;
 }
